Remove outdated files and components
The following files were removed due to being outdated or no longer in use:

- core/wolf_core/grafana_logger.py
- core/.github/workflows/pylint.yml
- core/wolf_core/runner.py
- core/docs/source/runner.rst
- core/.github/workflows/unittest.yml
- .github/workflows/deploy_sonu_dev_gzsim.yml
- wolf/install.py
- wolf/main.py
- .github/workflows/pylint.yml
- docs/source/runner.rst

This clean-up helps to declutter the development space and aligns the codebase with current progress and requirements.
