--- .github/workflows/unittest.yml
@@ -1,30 +1,0 @@
-name: Unittest
-
-on: [ push ]
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-    strategy:
-      matrix:
-        python-version: [ "3.10" ]
-
-    steps:
-      - uses: actions/checkout@v3
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v4
-        with:
-          python-version: ${{ matrix.python-version }}
-      - name: Install dependencies
-        run: |
-          python -m pip install --upgrade pip
-          python -m pip install virtualenv
-          python -m virtualenv venv
-          source venv/bin/activate
-          pip install poetry
-          poetry install
-      - name: Test with unittest
-        run: |
-          source venv/bin/activate
-          cd tests && python -m unittest
