+++ wolf_core/application.py
@@ -3,14 +3,15 @@
 This module contains the Application Interface of the core module.
 All applications must implement this interface if they want to be run by the runner.
 """
+import datetime
 import logging
+import os
 import threading
+import time
 import uuid
 from abc import ABC, abstractmethod
 from enum import Enum
 from typing import List
-
-import schedule
 
 from wolf_core import api
 
@@ -52,20 +53,67 @@ class Application(ABC):
     :see: :class:`api.API`
     :type _apis: List[api.API]
 
-    :ivar frequency: The frequency of the application.
-    :see: :class:`schedule.Job`
-    :type frequency: schedule.Job
     """
 
+    logger = logging.getLogger(__name__)
+    logger.setLevel(logging.DEBUG)
+
     def __init__(self):
+
         self._apis: List[api.API] = api.API.instances
-        self.frequency: schedule.Job = schedule.every().day
-        self.logger = logging.getLogger(__name__)
 
         self.__health_check = {"status": Status.WAITING, "message": " ", "last_execution": Status.NEVER, "id": -1, }
 
         self.app_lock = threading.Lock()
         self.__debug = False
+        self.__setup_logger()
+        self.__load_apis()
+
+    def __setup_logger(self):
+        """
+        This method sets up the logger.
+        It creates a file handler and a console handler.
+        The file handler logs all messages with level WARNING.
+
+        :return: None
+        """
+        self.logger.handlers = []
+        log_name = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
+
+        file_path = os.path.expanduser("~/.cache/wolf")
+        console_handler = logging.StreamHandler()
+        console_handler.setLevel(logging.DEBUG)
+        formatter = logging.Formatter('%(asctime)s - %(levelname)s - %(message)s')
+        console_handler.setFormatter(formatter)
+        self.logger.addHandler(console_handler)
+
+        try:
+            os.makedirs(os.path.join(file_path, 'log'), exist_ok=True)
+            file_handler = logging.FileHandler(file_path + "/log" + '/wolf_core_' + log_name + '.log')
+            file_handler.setLevel(logging.DEBUG)
+            file_handler.setFormatter(formatter)
+            self.logger.addHandler(file_handler)
+        except PermissionError:
+            self.logger.error("Could not create log file. Please check permissions.")
+
+        time.sleep(1)
+
+    def __load_apis(self):
+        """
+        This method loads the APIs by creating an instance of each API.
+        The instances are stored in the `_apis` list attribute.
+        It also sets the logs of each API.
+
+        :return: None
+        """
+        if len(self._apis) != 0:
+            self.logger.debug("Deleting old APIs.")
+            self._apis = []
+        for a in api.API.__subclasses__():
+            self._apis.append(a())
+            self.logger.debug("API " + a.__name__ + " loaded.")
+        for a in self._apis:
+            a.logger = self.logger
 
     @property
     def health_check(self):
